SOLID Single Responsibility principle
The reason here is that we area working in a class that is doing it responsibility that is setting up the services of the identity to the dependency injection collector. But in this single class we have the configuration for both presentation environment (WebAPI & WebAPP) and we area managing the configuration for authentication and authorization but I think it should be separated into different classes that each of them manage their security configurations.